+++ lib/instruction_decode/instruction_decode.ml
@@ -2,7 +2,14 @@ open Hardcaml
 open Hardcaml.Signal
 
 module I = struct
-  type 'a t = { instruction : 'a [@bits 32] } [@@deriving sexp_of, hardcaml]
+  type 'a t = {
+    clock : 'a;
+    writeback_reg_write_enable : 'a;
+    writeback_address : 'a; [@bits 5]
+    writeback_data : 'a; [@bits 32]
+    instruction : 'a; [@bits 32]
+  }
+  [@@deriving sexp_of, hardcaml]
 end
 
 module O = struct
@@ -16,20 +23,19 @@ module O = struct
   [@@deriving sexp_of, hardcaml]
 end
 
-let regfile rs rt =
-  (* We'll add write support when we get to the writeback stage. *)
+let delay_address_to_falling clock address = 
+  let spec = Reg_spec.create ~clock:clock () in
+  let spec_on_falling = Reg_spec.override ~clock_edge:Edge.Falling spec in
+  reg ~enable:vdd spec_on_falling address
+
+let regfile rs rt clock write_enable write_address write_data =
   let write_port =
-    {
-      write_clock = gnd;
-      write_address = (of_string "5'h1");
-      write_enable = gnd;
-      write_data = (of_string "32'h0");
-    }
+    { write_clock = clock; write_address; write_enable; write_data }
   in
   let number_of_regs = 32 in
+  let delay_address = delay_address_to_falling clock in
   let regs =
-    multiport_memory number_of_regs ~write_ports:[| write_port |]
-      ~read_addresses:[| rs; rt |]
+    multiport_memory number_of_regs ~write_ports:[|write_port|] ~read_addresses:[| delay_address rs; delay_address rt |]
   in
   (Array.get regs 0, Array.get regs 1)
 
@@ -38,7 +44,10 @@ let circuit_impl (scope : Scope.t) (input : _ I.t) =
     Control_unit.hierarchical scope { instruction = input.instruction }
   in
   let parsed = control_unit.parsed_instruction in
-  let rs_val, rt_val = regfile parsed.rs parsed.rt in
+  let rs_val, rt_val =
+    regfile parsed.rs parsed.rt input.clock input.writeback_reg_write_enable
+      input.writeback_address input.writeback_data
+  in
   {
     O.control_signals = control_unit.control_signals;
     rs_val;